+++ ModerneOpenGL/Source/src/Resources/ResourcesManager.cpp
@@ -1,6 +1,7 @@
 #include "ResourcesManager.h"
 #include "Model.h"
 #include "Log.h"
+
 ResourceManager* ResourceManager::instance = nullptr;
 
 ResourceManager::~ResourceManager()
@@ -60,7 +61,7 @@ void ResourceManager::Delete(std::string name)
 }
 std::string ResourceManager::GetExtension(std::string name)
 {
-    int index = name.find('.');
+    size_t index = name.find('.');
     std::string result = "";
 
     if (index < 0)
